+++ frontend/src/views/Simulator/RunDebugView.vue
@@ -193,99 +193,6 @@ const handleDeployContract = async ({
     />
 
     <div
-      class="flex flex-col overflow-y-auto"
-      v-if="!!contractsStore.currentContractId"
-    >
-      <!-- <div class="flex flex-col">
-        <div
-          class="flex w-full flex-col bg-slate-100 px-2 py-2 dark:bg-zinc-700"
-        >
-          <div class="text-sm">Intelligent Contract:</div>
-          <div
-            data-testid="current-contract-name"
-            class="text-xs text-neutral-800 dark:text-neutral-200"
-          >
-            {{ contractsStore.currentContract?.name }}
-          </div>
-        </div>
-
-        <ConstructorParameters
-          :inputs="contractsStore.currentConstructorInputs"
-          :loading="contractsStore.loadingConstructorInputs"
-          :error="contractsStore.currentErrorConstructorInputs"
-          @deploy-contract="handleDeployContract"
-          :deploying="contractsStore.deployingContract"
-        />
-      </div> -->
-
-      <div class="flex flex-col">
-        <!-- <div class="flex flex-col" v-show="contractsStore.deployedContract">
-          <ContractState
-            :abi="contractsStore.currentDeployedContractAbi"
-            :contract-state="contractsStore.currentContractState"
-            :deployed-contract="contractsStore.deployedContract"
-            :get-contract-state="handleGetContractState"
-            :calling-state="contractsStore.callingContractState"
-          />
-        </div> -->
-
-        <!-- <PageSection
-          v-if="
-            contractsStore.deployedContract &&
-            contractsStore.currentDeployedContractAbi?.methods
-          "
-        >
-          <template #title>Read Methods</template>
-          <ContractMethodItem
-            v-for="method in readMethods"
-            :key="method.methodName"
-            :methodName="method.methodName"
-            :method="method.method"
-            methodType="read"
-          />
-
-          <EmptyListPlaceholder v-if="readMethods.length === 0">
-            No read methods.
-          </EmptyListPlaceholder>
-        </PageSection> -->
-
-        <!-- <PageSection
-          v-if="
-            contractsStore.deployedContract &&
-            contractsStore.currentDeployedContractAbi?.methods
-          "
-        >
-          <template #title>Write Methods</template>
-
-          <ContractMethodItem
-            v-for="method in writeMethods"
-            :key="method.methodName"
-            :methodName="method.methodName"
-            :method="method.method"
-            methodType="write"
-          />
-
-          <EmptyListPlaceholder v-if="writeMethods.length === 0">
-            No write methods.
-          </EmptyListPlaceholder>
-        </PageSection> -->
-
-        <div class="flex flex-col" v-show="contractsStore.deployedContract">
-          <!-- <ExecuteTransactions
-            :abi="contractsStore.currentDeployedContractAbi"
-            @call-method="handleCallContractMethod"
-            :calling-method="contractsStore.callingContractMethod"
-          /> -->
-        </div>
-
-        <!-- <TransactionsList
-          :transactions="contractTransactions"
-          @clear-transactions="handleClearTransactions"
-        /> -->
-      </div>
-    </div>
-
-    <div
       class="flex w-full flex-col bg-slate-100 px-2 py-2 dark:dark:bg-zinc-700"
       v-else
     >
